pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "mt-butterfly"
version = "0.1.0"
description = "Web service for chatting with OpenCode"
readme = "README.md"
requires-python = ">=3.10"
authors = [{name = "Lucha"}]
dependencies = [
    "fastapi==0.115.5",
    "pydantic-settings==2.6.1",
    "uvicorn[standard]==0.32.1",
    "jinja2==3.1.4",
    "python-multipart==0.0.12",
    "sqlalchemy==2.0.36",
    "aiosqlite==0.20.0",
    "apscheduler==3.10.4",
    "python-dotenv==1.0.1",
    "httpx==0.28.0",
    "pytest==8.3.4",
    "pytest-asyncio==0.24.0",
    "anyio==4.7.0",
    "websockets==14.1",
    "greenlet==3.1.1",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
]

[project.scripts]
mt-butterfly = "app.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.setuptools.package-data]
app = ["templates/**/*", "static/**/*"]
